added withRnvRNConfig, added missing files to template
