+++ packages/engine-rn-macos/src/adapters/metroAdapter.ts
@@ -75,22 +75,3 @@ export const withRNVMetro = (config: any) => {
 
     return cnf;
 };
-
-export const withRNVBabel = (cnf: any) => {
-    const plugins = cnf?.plugins || [];
-
-    return {
-        retainLines: true,
-        presets: ['module:metro-react-native-babel-preset'],
-        ...cnf,
-        plugins: [
-            [
-                require.resolve('babel-plugin-module-resolver'),
-                {
-                    root: [process.env.RNV_MONO_ROOT || '.'],
-                },
-            ],
-            ...plugins,
-        ],
-    };
-};
