+++ packages/engine-rn-tvos/src/index.ts
@@ -16,6 +16,7 @@ import taskRnvLog from './tasks/task.rnv.log';
 import CNF from '../renative.engine.json';
 import { withRNVBabel } from './adapters/babelAdapter';
 import { withRNVMetro } from './adapters/metroAdapter';
+import { withRNVRNConfig } from './adapters/rnConfigAdapter';
 
 const Engine: RnvEngine = {
     // initializeRuntimeConfig: (c) => Context.initializeConfig(c),
@@ -68,6 +69,6 @@ const Engine: RnvEngine = {
 
 const withRNV = withRNVMetro;
 
-export { withRNVMetro, withRNV, withRNVBabel };
+export { withRNVMetro, withRNV, withRNVBabel, withRNVRNConfig };
 
 export default Engine;
