+++ packages/engine-rn-windows/src/adapters/metroAdapter.ts
@@ -83,21 +83,3 @@ export const withRNVMetro = (config: any) => {
     return cnf;
 };
 
-export const withRNVBabel = (cnf: any) => {
-    const plugins = cnf?.plugins || [];
-
-    return {
-        retainLines: true,
-        presets: ['module:metro-react-native-babel-preset'],
-        ...cnf,
-        plugins: [
-            [
-                require.resolve('babel-plugin-module-resolver'),
-                {
-                    root: [process.env.RNV_MONO_ROOT || '.'],
-                },
-            ],
-            ...plugins,
-        ],
-    };
-};
